Separacion de archivos por clases
