+++ main.py
@@ -1,36 +1,7 @@
-from textblob import TextBlob
-from googletrans import Translator
+from sentimiento_analizer import SentimientoAnalizer
+from text_processor import TextProcessor
+from translator_service import TranslatorService 
 
-
-class TranslatorService:
-    def traduccion_texto(self, texto: str) -> str:
-        traduccion = Translator()
-        #print(texto_traducido)
-        return traduccion.translate(texto, src='es', dest='en').text
-
-
-class SentimientoAnalizer:
-    def analizar_sentimiento(self, texto_traducido: str) -> str:
-        blob = TextBlob(texto_traducido)
-        sentimiento = blob.sentiment.polarity
-        #print('Polaridad: ', sentimiento)
-
-        if sentimiento > 0:
-            return 'Positivo'
-        elif sentimiento < 0:
-            return 'Negativo'
-        else:
-            return 'Neutral'
-
-class TextProcessor:
-    def __init__(self, translator:TranslatorService, analyzer:SentimientoAnalizer):
-        self.translator = translator
-        self.analyzer = analyzer
-
-    def process(self, texto: str) -> str:
-        texto_traducido = self.translator.traduccion_texto(texto)
-        sentimiento = self.analyzer.analizar_sentimiento(texto_traducido)
-        return sentimiento    
 
 def run():
     texto = input('Ingrese un texto para revisar: ')
